requirements.txt
BeautifulSoup
pygments
django
djangorestframework
markdown
django-filter
django_summernote
django-bootstrap3
django-fontawesome
django-datatable
django-ajax-selects